% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeSettings.R
\name{changeSettings}
\alias{changeSettings}
\title{Change user settings}
\usage{
changeSettings(settings)
}
\arguments{
\item{settings}{a named list - whose keys become separate keyword arguments
and the values become values of these arguments
ex: changeSettings(list(age_filter_min=30, gender=0))
age_filter_min: 18..46
age_filter_max: 22..55
age_filter_min <= age_filter_max - 4
gender: 0 == seeking males, 1 == seeking females
distance_filter: 1..100
discoverable: true | false
{"photo_optimizer_enabled":false}}
}
\description{
\code{changeSettings} Change own user settings
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}
