% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configAccount.R
\name{configAccount}
\alias{configAccount}
\title{Configure tinder account}
\usage{
configAccount(mail, password, saveAccountFile = paste0(Sys.getenv("HOME"),
  "/.tinderAccount.R"))
}
\arguments{
\item{mail}{mail to login in Facebook.}

\item{password}{password to login in Facebook.}

\item{saveAccountFile}{string with the file where to save account info.}
}
\description{
\code{configAccount} Authorizes in Facebook to use tinder App.
Saves corresponding Facebook ID and access token to file.
Has to be done once. After, \code{login} will continue using this info.
}
\examples{
\dontrun{
# When first using Rtinder, we should configure our account.
# It will save (by default) the config file in ~/.tinderAccount.R .
configAccount('MY_FB_REGISTERED_MAIL', 'MY_FB_PASSWORD');
}

}
\seealso{
{login}
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}
