% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchInfo.R
\name{getMatchInfo}
\alias{getMatchInfo}
\title{Match info}
\usage{
getMatchInfo(matchId)
}
\arguments{
\item{matchId}{string containing the match id}
}
\description{
\code{getMatchInfo} Get match information
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}
