% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOwnMetadata.R
\name{getOwnMetadata}
\alias{getOwnMetadata}
\title{Get own user metadata}
\usage{
getOwnMetadata()
}
\description{
\code{getOwnMetadata} Returns metadata on yourself. Including the following keys:
['globals', 'client_resources', 'versions', 'purchases',
 'status', 'groups', 'products', 'rating', 'tutorials',
  'travel', 'notifications', 'user']
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}

