% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOwnProfile.R, R/getProfile.R
\name{getOwnProfile}
\alias{getOwnProfile}
\alias{getProfile}
\title{Get own profile data}
\usage{
getOwnProfile()

getProfile(id)
}
\arguments{
\item{id}{string containing desired user id}
}
\description{
\code{getOwnProfile} Returns your own profile data

\code{getOwnProfile} Gets a user's profile via their id
}
