% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdates.R
\name{getUpdates}
\alias{getUpdates}
\title{Get updates}
\usage{
getUpdates(lastActivityDate = "")
}
\arguments{
\item{lastActivityDate}{The last activity date is defaulted at the beginning of time.
Format for lastActivityDate: "2017-07-09T10:28:13.392Z"}
}
\description{
\code{getUpdates} Returns all updates since the given activity date.
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}

