% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.R
\name{like}
\alias{dislike}
\alias{like}
\alias{reportPerson}
\alias{superlike}
\title{Likes}
\usage{
like(personId)

superlike(personId)

dislike(personId)

reportPerson(personId, cause, explanation = "")
}
\arguments{
\item{personId}{string containing the desired person ID}

\item{cause}{string, there are three options for cause:
0 : Other and requires an explanation
1 : Feels like spam and no explanation
4 : Inappropriate Photos and no explanation}

\item{explanation}{string with the reporting explanation.?}
}
\description{
\code{like} Likes a user.
\code{superlike} Superlikes a user.
\code{dislike} Dislikes a user.
\code{reportPerson} Reports a user.
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}

