% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{changeLocation}
\alias{changeLocation}
\alias{resetLocation}
\title{Modify user location}
\usage{
changeLocation(lat, lon)
}
\arguments{
\item{lat}{string with the latitude}

\item{lon}{string with the longitude}
}
\description{
\code{changeLocation} Updates your location to the given float inputs
Note: Careful!
\code{resetLocation} Resets your location (to GPS)
\code{changeLocationPlus} Updates your location to the given float inputs
Note: Requires a passport / Tinder Plus
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}
