% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to tinder account}
\usage{
login(accountFile = paste0(Sys.getenv("HOME"), "/.tinderAccount.R"))
}
\arguments{
\item{accountFile}{R file to be sourced, with 'userId' and 'fbAuthtoken'
variables.}
}
\description{
\code{login} Loads tinder's facebook access id and token. And checks if can
connect.
}
\details{
Sources the selected R file, which must have 'userId' and 'accesstoken'
variables. And tests if it can connect to tinder's API.
}
\examples{
\dontrun{
# Check if can connect to tinder.
# login() must be executed each time we load Rtinder library.
login();
}

}
\seealso{
{configAccount}
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}
