% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendMessage.R
\name{sendMessage}
\alias{sendMessage}
\title{Send message}
\usage{
sendMessage(matchId, message)
}
\arguments{
\item{matchId}{string with the ID of the match}

\item{message}{string with the message to send}
}
\description{
\code{sendMessage} Sends a message to a match.
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}

