% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webUsername.R
\name{setWebUsername}
\alias{resetWebUsername}
\alias{setWebUsername}
\title{Modify owns username for the webprofile}
\usage{
setWebUsername(username)
}
\arguments{
\item{username}{string with the desired username}
}
\description{
\code{setWebUsername} sets the username for the webprofile: https://www.gotinder.com/@YOURUSERNAME
\code{resetWebUsername} resets the username for the webprofile: https://www.gotinder.com/@YOURUSERNAME
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}

